+++ py/just_mospec/test/__init__.py
@@ -1,7 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE
 # -*- coding: utf-8 -*-
 """
-justmospec.test
+just_mospec.test
 ===============
 
 Used to initialize the unit test framework via ``pytest``.